requirements.txt
flask
flask-cors
python-dotenv
openai
transformers
chromadb
gTTS
torch>=2.6